Android Sunflower: Java to Kt, grow zones basics
Code reviewers: due to the shift of Kotlin becoming the preferred
Android development language, this CL is mostly comprised of converting
Java files to Kotlin. A minimal set of Java files (mostly around
PlantDetail) will remain in Java to illustrate interop.

This CL also adds the basics for associating a plant with a grow zone.
The zone represents where the plant would thrive the best and is a
property that is used for filtering the list of plants shown in the
RecyclerView.  There is a very basic UI around filtering, which is a WIP
that will be replaced with a proper UI in a subsequent CL,
in order to keep this CL small.

This is step 8 in the project tracker.

Other changes include:
- Minor build.gradle cleanup
- Add Tests:
  - Util for testing LiveData objects
  - PlantDao
- Change the following from Java to Kotlin:
  - AppDatabase
  - ExampleInstrumentedTest
  - PlantRepository
  - PlantListActivity
  - PlantListActivityTest
  - PlantTest
  - TestUtils

Tracker: go/android-sunflower-tracker
PRD: go/android-sunflower-prd

Bug: 65952046
Test: ran unit and instrumented tests:
./gradlew app:test app:connectedAndroidTest

Change-Id: If241a6f3526f88a4117dd28831b58c47ce1fc73f