Android Sunflower: Add watering interval to Plant
Adds wateringInterval to Plant, which is how often the plant should be
watered, in days.  Add function to determine if a plant should be
watered, given a previous watering day.

This is some pre-work for showing notifications to prompt the user to
water the plant(s) in their garden.

Tracker: go/android-sunflower-tracker
PRD: go/android-sunflower-prd

Bug: 65952046
Test: ran unit and instrumented tests:
./gradlew app:test app:connectedAndroidTest

Change-Id: Ib2eb3f116a4f70a4cae4cffddecd1ea9082cd220