Update failing test.
Don't check that PlantAndGardenPlantings.plant is null when
build with the default constructor because it should always
be set by Room to a valid plant (since it represents a plant
in a garden).

The fact that it is non-null (and points to the correct plant)
is tested in GardenPlantingDaoTest.testGetPlantAndGardenPlantings.