Upgrade from kapt to ksp and update dependencies
Issue-917: Introduced the ksp plugin in project dagger and hilt dependencies now use ksp instead of kapt. Removed kapt plugin and use overall in project. Changed kotlin version from '1.8.0' to '1.9.10'. Changed hilt version from 2.44.2 to 2.48 as it is the version that supports ksp. Had to change compose-compiler version to 1.5.3 as a result of upgrading kotlin version to '1.9.10'.