Update README.md to deprecate Sunflower
We are prioritizing a smaller set of Compose samples.

Please see https://github.com/android/compose-samples for a more up-to-date set of Compose showcases. 

We encourage you to fork this sample if you are still finding it valuable for your use case, however - no more updates will be made to this repository.